splitting off members (begin)
